Navodila in  python skripta za testne primere

+++ Domače naloge/DN02 - Zlivanje/Marjan.sh
@@ -1,5 +1,5 @@
 #####################################
-			VERSION=1.5
+			VERSION=1.7.2
 #####################################
 
 #./Marjan.sh {ime_datoteke} {dodatni_parametri}
@@ -20,19 +20,36 @@ reset='\e[0m'
 # Preveri, če je podanih dovolj argumentov (ime programa)
 if [ "$#" -eq 0 ]; then
   echo "Uporaba: $0 <ime_programa> <dodatni_parametri>"
-  echo "Dodatni parametri: T={} - timeoutlimit, M={} - measuretime, R={} - range of tests (1- || -4 || 2-5)"
+  echo "Dodatni parametri: T={} - timeoutlimit, M={} - measuretime (0 - NE, 1 - DA), R={} - range of tests (1- || -4 || 2-5)"
   exit 1
 fi
 
 # Argumenti
 program="$1"
 shift 1
+remove=0
+
+if [ ! -f "${program}" ]; then
+echo "Datoteka ${program} in ${program}.cpp ne obstaja."
+exit 1
+else
+remove=0
+# če je datoteka s končnico .cpp
+if [[ "${program}" == *".cpp" ]]; then
+	program="${program%.*}"
+	g++ "${program}.cpp" -o ${program} -std=c++20 -pedantic -Wall
+	remove=1
+	fi
+fi
+
+
+
 
 # Čas za timeout
-to=2.0
+to=2.5
 
 # Izpiši čas
-M_time=0
+M_time=1
 total_time=0
 
 # Testno območje
@@ -87,7 +104,8 @@ echo -e "${white}Measure time = ${M_time}${reset}"
 echo -e "${white}Test range = ${range}${reset}"
 echo ""
 
-g++ "${program}.cpp" -o ${program} -std=c++20 -pedantic -Wall
+
+
 
 # Pravilni testi
 correct_tests=0
@@ -100,6 +118,7 @@ test_count=0
 
 # funkcija ob koncu izvajanja
 izpisi_rezultate() {
+	preimenuj=0
 	# Izpiši število pravilnih testov na koncu
 	echo -e "${white}---------------"
 	if [ $((test_count -  done_tests)) -ne 0 ]; then
@@ -107,6 +126,7 @@ izpisi_rezultate() {
 	fi
 	if [ $done_tests -ne 0 ]; then
 	  echo "Ostali opravljeni testi : $done_tests"
+	  preimenuj=1
 	fi
 
 	if [[ ${M_time} = 1 ]]; then
@@ -118,10 +138,34 @@ izpisi_rezultate() {
 	  averagetime="$averagetime s"
 	  echo "Average time = ${averagetime}"
 	fi
-
+	
+	if [ $preimenuj -eq 1 ]; then
+	  echo ""
+	  echo -e "${blue}Do you want to rename .res to .out? (y/n)${reset}"
+	  read -r answer
+	  if [[ $answer =~ ^[Yy]$ ]]; then
+	    echo -e "${white}Renaming..."
+		for input_file in test*.in;
+		do
+		  input_number=$(echo "$input_file" | sed 's/test\([0-9]\{2\}\)\.in/\1/')
+		  output_file_rez="test$input_number.out"
+		  output_file_temp="test$input_number.res"
+		  if [ -f $output_file_rez ]; then
+			continue
+		  else
+		    if [ -f $output_file_temp ]; then
+		  	  mv $output_file_temp $output_file_rez
+			fi  
+		  fi
+		done
+		echo -e "${green}Done."
+	  fi  	
+	fi
+	
 	echo -e ${reset}
-
-	rm "${program}"
+	if [ $remove -eq 1 ]; then
+	  rm "${program}"
+	fi
 	exit 1
 }
 
@@ -179,6 +223,12 @@ testing() {
 	  # Preveri, ali je program presegel časovni limit
 	  if [ $rezultat -eq 124 ]; then
 		echo -e "Test $input_number: ${purple}Timeout${reset}${measuredTimeString}"
+		rm $output_file_temp
+		if [ -f $output_file_rez ];then
+		  rezultat=-1
+		else
+		   done_tests=$((done_tests + 1))
+		fi
 	  else
 		if [ -f $output_file_rez ]; then
 			# Primerjaj izhod programa s pričakovanim izhodom
